+++ lib/eps/base_estimator.rb
@@ -1,6 +1,9 @@
 module Eps
   class BaseEstimator
     def initialize(data = nil, y = nil, **options)
+      @options = options.dup
+      # TODO better pattern - don't pass most options to train
+      options.delete(:intercept)
       train(data, y, **options) if data
     end