+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## 0.3.1 [unreleased]
+## 0.3.1 (unreleased)
 
 - Added `weight` option for LightGBM and linear regression
+- Added `intercept` option for linear regression
 
 ## 0.3.0 (2019-09-05)